feat(entities): redesign Entity scripts to work with the state machine
The previous Entity approach was based on using components that were
responsible for their own area - damage, health, hitboxes, movement, etc.

After a playtest of the `r2` version, it turned out that this organization does
not allow to separately distinguish the "states" of an entity at a particular
point in time. Also, over time, the logic of the components is becoming
more complex and confusing.

The new organization is based on the use of states, in which execution
conditions, logic, and the corresponding animation are clearly defined.

**NOTE**: the preparatory code for a future implementation of save/load
has been removed and will probably be implemented later.